Update DisplayControl
Added support for updated OBS versions with new window class type. Taking a page from LDA, this now reads in an ini file for the OBS window class type. If OBS is not found, a popup window appears to let you select OBS. This class type is then saved out to the ini for future runs of the program.
Included Acc due to VEX_Shared requiring it